# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Toolstead Platform API
  version: 0.4.0
  description: Workspace, module entitlement, and CRM foundation for Toolstead.
servers:
  - url: /api/v1
security:
  - accessCookie: []
paths:
...
  /contacts:
    get:
      operationId: listContacts
      parameters:
        - name: q
          in: query
          schema:
            type: string
            maxLength: 120
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: Active workspace contacts.
    post:
      operationId: createContact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "201":
          description: Contact created.
        "400":
          $ref: "#/components/responses/BadRequest"
  /contacts/{contactId}:
    patch:
      operationId: updateContact
      parameters:
        - $ref: "#/components/parameters/ContactId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "200":
          description: Contact updated.
    delete:
      operationId: archiveContact
      parameters:
        - $ref: "#/components/parameters/ContactId"
      responses:
        "204":
          description: Contact archived.
...
  /contacts/{contactId}/notes:
    post:
      operationId: addContactNote
      parameters:
        - $ref: "#/components/parameters/ContactId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [body]
              properties:
                body:
                  type: string
                  minLength: 1
                  maxLength: 2000
      responses:
        "201":
          description: Timeline note created.
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    accessCookie:
      type: apiKey
      in: cookie
      name: toolstead_access
    refreshCookie:
      type: apiKey
      in: cookie
      name: toolstead_refresh
  parameters:
    ContactId:
      name: contactId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    Forbidden:
      description: Role or module entitlement denied.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    ContactInput:
      type: object
      additionalProperties: false
      required: [displayName]
      properties:
        displayName:
          type: string
          minLength: 1
          maxLength: 180
        companyName:
          type: string
          maxLength: 180
        email:
          type: string
          format: email
          maxLength: 320
        phone:
          type: string
          maxLength: 40
        source:
          type: string
          maxLength: 80
        stage:
          type: string
          enum: [New lead, Contacted, Qualified, Won, Lost]
        summary:
          type: string
          maxLength: 2000
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code:
              type: string
            message:
              type: string
            requestId:
              type: string